+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/CallbackQuery/MessageCallbackQuery.kt
@@ -1,7 +1,8 @@
 package dev.inmo.tgbotapi.types.CallbackQuery
 
-import dev.inmo.tgbotapi.types.message.abstracts.Message
+import dev.inmo.tgbotapi.types.message.abstracts.ContentMessage
+import dev.inmo.tgbotapi.types.message.content.abstracts.MessageContent
 
 sealed interface MessageCallbackQuery : CallbackQuery {
-    val message: Message
+    val message: ContentMessage<MessageContent>
 }