+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/CallbackQuery/MessageDataCallbackQuery.kt
@@ -2,12 +2,13 @@ package dev.inmo.tgbotapi.types.CallbackQuery
 
 import dev.inmo.tgbotapi.types.CallbackQueryIdentifier
 import dev.inmo.tgbotapi.types.User
-import dev.inmo.tgbotapi.types.message.abstracts.Message
+import dev.inmo.tgbotapi.types.message.abstracts.ContentMessage
+import dev.inmo.tgbotapi.types.message.content.abstracts.MessageContent
 
 data class MessageDataCallbackQuery(
     override val id: CallbackQueryIdentifier,
     override val from: User,
     override val chatInstance: String,
-    override val message: Message,
+    override val message: ContentMessage<MessageContent>,
     override val data: String
 ) : DataCallbackQuery, MessageCallbackQuery