+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/ChatEvents/UserLoggedIn.kt
@@ -1,0 +1,7 @@
+package dev.inmo.tgbotapi.types.message.ChatEvents
+
+import dev.inmo.tgbotapi.types.message.ChatEvents.abstracts.PrivateEvent
+
+data class UserLoggedIn(
+    val domain: String
+) : PrivateEvent