+++ tgbotapi.utils/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/utils/extensions/raw/PreCheckoutQuery.kt
@@ -1,0 +1,20 @@
+package dev.inmo.tgbotapi.extensions.utils.extensions.raw
+
+import dev.inmo.tgbotapi.types.InvoicePayload
+import dev.inmo.tgbotapi.types.ShippingOptionIdentifier
+import dev.inmo.tgbotapi.types.payments.OrderInfo
+import dev.inmo.tgbotapi.types.payments.PreCheckoutQuery
+import dev.inmo.tgbotapi.utils.RiskFeature
+
+@RiskFeature(RawFieldsUsageWarning)
+val PreCheckoutQuery.total_amount: Long
+    get() = amount
+@RiskFeature(RawFieldsUsageWarning)
+val PreCheckoutQuery.invoice_payload: InvoicePayload
+    get() = invoicePayload
+@RiskFeature(RawFieldsUsageWarning)
+val PreCheckoutQuery.shipping_option_id: ShippingOptionIdentifier?
+    get() = shippingOptionId
+@RiskFeature(RawFieldsUsageWarning)
+val PreCheckoutQuery.order_info: OrderInfo?
+    get() = orderInfo