+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.38.1
 
+* `Core`:
+    * `MessageCallbackQuery` (and all implementers as well) has changed the type of `message`: now it is `ContentMessage<MessageContent>` instead of `Message`
+
 ## 0.38.0
 
 _This update contains [Telegram Bot API 5.6](https://core.telegram.org/bots/api-changelog#december-30-2021) implementation_